Move test to ServerBasicTests
